refactor: 拆出選取規則、鍵盤範圍與資料夾列，sidebar_panel 降到 800 行以下
1. branch_selection_rules.dart 收攏純規則：isBulkSelectable、isInMultiSelection、
   isGoneAndBulkSelectable、prunedSelection、extendedSelection、liveBranchNames。
   面板原本有四處各自用略不同的說法問「這一列可不可以選」
2. firstLeafName 與 selectableLeafNames 是對樹的純走訪，移到
   branch_tree_builder.dart，與 collectFolderLeafRefs 同一個理由。呼叫點仍留在
   build() 走同一棵 branchTree，渲染順序不會出現第二份
3. BranchSelectionShortcuts 與 BranchFolderRow 各自成檔
4. sidebar_panel.dart 965 → 741 行；features/sidebar 底下每個檔案都在 800 行以內
5. 全程無行為變動，1988 個測試逐步保持全綠
